Use integrated embeddings
Currently I use embeddings which can be done internally in pinecone.
This means instead of creating the vectors myself I can just upload the
data and let pinecone handle this.

This makes the setup easier and removes one of the longest running task
locally (mainly for creating the whole index).