Split evaluation and ai tool from each other
Now I can quickly change what tools to use and additionally have better
comparison, as the evaluation does not get unusable if the model itself
just gives bad feedback. Which would result in results that can not be
trusted.

With this a comparison between tools is possible, as the evaulation tool
can stay the same, resulting in somewhat valid results.